AbbrSettings: Replaced pykdeuic4 generated code with cleaned up original ui file. Moved AbbrSettings and AbbrListItem classes from dialogs.py to abbrsettings.py.
