+++ lib/autokey/qtui/abbrsettings.py
@@ -1,105 +1,95 @@
-#!/usr/bin/env python
-# coding=UTF-8
-#
-# Generated by pykdeuic4 from abbrsettings.ui on Sun Mar  4 11:39:39 2012
-#
-# WARNING! All changes to this file will be lost.
-from PyKDE4 import kdecore
-from PyKDE4 import kdeui
-from PyQt4 import QtCore, QtGui
+# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2018 Thomas Hess <thomas.hess@udo.edu>
 
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    _fromUtf8 = lambda s: s
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 
-class Ui_Form(object):
-    def setupUi(self, Form):
-        Form.setObjectName(_fromUtf8("Form"))
-        Form.resize(571, 350)
-        self.verticalLayout_3 = QtGui.QVBoxLayout(Form)
-        self.verticalLayout_3.setObjectName(_fromUtf8("verticalLayout_3"))
-        self.horizontalLayout_3 = QtGui.QHBoxLayout()
-        self.horizontalLayout_3.setObjectName(_fromUtf8("horizontalLayout_3"))
-        self.verticalLayout_2 = QtGui.QVBoxLayout()
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.abbrListWidget = QtGui.QListWidget(Form)
-        sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Ignored, QtGui.QSizePolicy.Expanding)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.abbrListWidget.sizePolicy().hasHeightForWidth())
-        self.abbrListWidget.setSizePolicy(sizePolicy)
-        self.abbrListWidget.setAlternatingRowColors(True)
-        self.abbrListWidget.setUniformItemSizes(True)
-        self.abbrListWidget.setObjectName(_fromUtf8("abbrListWidget"))
-        self.verticalLayout_2.addWidget(self.abbrListWidget)
-        self.horizontalLayout_2 = QtGui.QHBoxLayout()
-        self.horizontalLayout_2.setObjectName(_fromUtf8("horizontalLayout_2"))
-        self.addButton = QtGui.QPushButton(Form)
-        icon = QtGui.QIcon()
-        icon.addPixmap(QtGui.QPixmap(), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.addButton.setIcon(icon)
-        self.addButton.setFlat(False)
-        self.addButton.setObjectName(_fromUtf8("addButton"))
-        self.horizontalLayout_2.addWidget(self.addButton)
-        self.removeButton = QtGui.QPushButton(Form)
-        self.removeButton.setText(_fromUtf8(""))
-        self.removeButton.setIcon(icon)
-        self.removeButton.setObjectName(_fromUtf8("removeButton"))
-        self.horizontalLayout_2.addWidget(self.removeButton)
-        self.verticalLayout_2.addLayout(self.horizontalLayout_2)
-        self.horizontalLayout_3.addLayout(self.verticalLayout_2)
-        self.verticalLayout = QtGui.QVBoxLayout()
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.horizontalLayout = QtGui.QHBoxLayout()
-        self.horizontalLayout.setObjectName(_fromUtf8("horizontalLayout"))
-        self.triggerOnLabel = QtGui.QLabel(Form)
-        self.triggerOnLabel.setObjectName(_fromUtf8("triggerOnLabel"))
-        self.horizontalLayout.addWidget(self.triggerOnLabel)
-        self.wordCharCombo = KComboBox(Form)
-        self.wordCharCombo.setEditable(True)
-        self.wordCharCombo.setObjectName(_fromUtf8("wordCharCombo"))
-        self.horizontalLayout.addWidget(self.wordCharCombo)
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        self.verticalLayout.addLayout(self.horizontalLayout)
-        self.removeTypedCheckbox = QtGui.QCheckBox(Form)
-        self.removeTypedCheckbox.setObjectName(_fromUtf8("removeTypedCheckbox"))
-        self.verticalLayout.addWidget(self.removeTypedCheckbox)
-        self.omitTriggerCheckbox = QtGui.QCheckBox(Form)
-        self.omitTriggerCheckbox.setObjectName(_fromUtf8("omitTriggerCheckbox"))
-        self.verticalLayout.addWidget(self.omitTriggerCheckbox)
-        self.matchCaseCheckbox = QtGui.QCheckBox(Form)
-        self.matchCaseCheckbox.setObjectName(_fromUtf8("matchCaseCheckbox"))
-        self.verticalLayout.addWidget(self.matchCaseCheckbox)
-        self.ignoreCaseCheckbox = QtGui.QCheckBox(Form)
-        self.ignoreCaseCheckbox.setObjectName(_fromUtf8("ignoreCaseCheckbox"))
-        self.verticalLayout.addWidget(self.ignoreCaseCheckbox)
-        self.triggerInsideCheckbox = QtGui.QCheckBox(Form)
-        self.triggerInsideCheckbox.setObjectName(_fromUtf8("triggerInsideCheckbox"))
-        self.verticalLayout.addWidget(self.triggerInsideCheckbox)
-        self.immediateCheckbox = QtGui.QCheckBox(Form)
-        self.immediateCheckbox.setObjectName(_fromUtf8("immediateCheckbox"))
-        self.verticalLayout.addWidget(self.immediateCheckbox)
-        spacerItem1 = QtGui.QSpacerItem(20, 40, QtGui.QSizePolicy.Minimum, QtGui.QSizePolicy.Expanding)
-        self.verticalLayout.addItem(spacerItem1)
-        self.horizontalLayout_3.addLayout(self.verticalLayout)
-        self.verticalLayout_3.addLayout(self.horizontalLayout_3)
-        self.kseparator = KSeparator(Form)
-        self.kseparator.setObjectName(_fromUtf8("kseparator"))
-        self.verticalLayout_3.addWidget(self.kseparator)
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
-        self.retranslateUi(Form)
-        QtCore.QMetaObject.connectSlotsByName(Form)
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-    def retranslateUi(self, Form):
-        Form.setWindowTitle(kdecore.i18n(_fromUtf8("Form")))
-        self.triggerOnLabel.setText(kdecore.i18n(_fromUtf8("Trigger on:")))
-        self.removeTypedCheckbox.setText(kdecore.i18n(_fromUtf8("Remove typed abbreviation")))
-        self.omitTriggerCheckbox.setText(kdecore.i18n(_fromUtf8("Omit trigger character")))
-        self.matchCaseCheckbox.setText(kdecore.i18n(_fromUtf8("Match phrase case to typed abbreviation")))
-        self.ignoreCaseCheckbox.setText(kdecore.i18n(_fromUtf8("Ignore case of typed abbreviation")))
-        self.triggerInsideCheckbox.setText(kdecore.i18n(_fromUtf8("Trigger when typed as part of a word")))
-        self.immediateCheckbox.setText(kdecore.i18n(_fromUtf8("Trigger immediately (don\'t require a trigger character)")))
+from PyQt4 import uic
+from PyQt4 import QtCore, Qt
+from PyQt4.QtGui import QListWidgetItem
 
-from PyKDE4.kdeui import KSeparator, KComboBox
+
+from .common import get_ui_qfile, EMPTY_FIELD_REGEX
+
+ui_file = get_ui_qfile("abbrsettings")
+AbbrSettingsBase = uic.loadUiType(ui_file)
+ui_file.close()
+
+
+WORD_CHAR_OPTIONS_ORDERED = ["All non-word", "Space and Enter", "Tab"]
+
+class AbbrListItem(QListWidgetItem):
+
+    def __init__(self, text):
+        QListWidgetItem.__init__(self, text)
+        self.setFlags(self.flags() | Qt.ItemFlags(Qt.ItemIsEditable))
+
+    def setData(self, role, value):
+        if value == "":
+            self.listWidget().itemChanged.emit(self)
+        else:
+            QListWidgetItem.setData(self, role, value)
+
+
+class AbbrSettings(*AbbrSettingsBase):
+
+    def __init__(self, parent):
+        super().__init__(parent)
+        self.setupUi()
+        for item in WORD_CHAR_OPTIONS_ORDERED:
+            self.wordCharCombo.addItem(item)
+
+    def setupUi(self):
+        self.setObjectName("Form")
+        super().setupUi(self)
+        QtCore.QMetaObject.connectSlotsByName(self)
+
+    def on_addButton_pressed(self):
+        item = AbbrListItem("")
+        self.abbrListWidget.addItem(item)
+        self.abbrListWidget.editItem(item)
+        self.removeButton.setEnabled(True)
+
+    def on_removeButton_pressed(self):
+        item = self.abbrListWidget.takeItem(self.abbrListWidget.currentRow())
+        if self.abbrListWidget.count() == 0:
+            self.removeButton.setEnabled(False)
+
+    def on_abbrListWidget_itemChanged(self, item):
+        if EMPTY_FIELD_REGEX.match(item.text()):
+            row = self.abbrListWidget.row(item)
+            self.abbrListWidget.takeItem(row)
+            del item
+
+        if self.abbrListWidget.count() == 0:
+            self.removeButton.setEnabled(False)
+
+    def on_abbrListWidget_itemDoubleClicked(self, item):
+        self.abbrListWidget.editItem(item)
+
+    def on_ignoreCaseCheckbox_stateChanged(self, state):
+        if not self.ignoreCaseCheckbox.isChecked():
+            self.matchCaseCheckbox.setChecked(False)
+
+    def on_matchCaseCheckbox_stateChanged(self, state):
+        if self.matchCaseCheckbox.isChecked():
+            self.ignoreCaseCheckbox.setChecked(True)
+
+    def on_immediateCheckbox_stateChanged(self, state):
+        if self.immediateCheckbox.isChecked():
+            self.omitTriggerCheckbox.setChecked(False)
+            self.omitTriggerCheckbox.setEnabled(False)
+            self.wordCharCombo.setEnabled(False)
+        else:
+            self.omitTriggerCheckbox.setEnabled(True)
+            self.wordCharCombo.setEnabled(True)
