+++ lib/autokey/qtui/dialogs.py
@@ -35,9 +35,8 @@ WORD_CHAR_OPTIONS = {
                      "Space and Enter" : r"[^ \n]",
                      "Tab" : r"[^\t]"
                      }
-WORD_CHAR_OPTIONS_ORDERED = ["All non-word", "Space and Enter", "Tab"]
 
-EMPTY_FIELD_REGEX = re.compile(r"^ *$", re.UNICODE)
+from .common import EMPTY_FIELD_REGEX
 
 def validate(expression, message, widget, parent):
     if not expression:
@@ -46,70 +45,11 @@ def validate(expression, message, widget, parent):
             widget.setFocus()
     return expression
 
-class AbbrListItem(QListWidgetItem):
 
-    def __init__(self, text):
-        QListWidgetItem.__init__(self, text)
-        self.setFlags(self.flags() | Qt.ItemFlags(Qt.ItemIsEditable))
 
-    def setData(self, role, value):
-        if value == "":
-            self.listWidget().itemChanged.emit(self)
-        else:
-            QListWidgetItem.setData(self, role, value)
+AbbrSettings = abbrsettings.AbbrSettings
+WORD_CHAR_OPTIONS_ORDERED = abbrsettings.WORD_CHAR_OPTIONS_ORDERED
 
-class AbbrSettings(QWidget, abbrsettings.Ui_Form):
-
-    def __init__(self, parent):
-        QWidget.__init__(self, parent)
-        abbrsettings.Ui_Form.__init__(self)
-        self.setupUi(self)
-
-        for item in WORD_CHAR_OPTIONS_ORDERED:
-            self.wordCharCombo.addItem(item)
-
-        self.addButton.setIcon(KIcon("list-add"))
-        self.removeButton.setIcon(KIcon("list-remove"))
-
-    def on_addButton_pressed(self):
-        item = AbbrListItem("")
-        self.abbrListWidget.addItem(item)
-        self.abbrListWidget.editItem(item)
-        self.removeButton.setEnabled(True)
-
-    def on_removeButton_pressed(self):
-        item = self.abbrListWidget.takeItem(self.abbrListWidget.currentRow())
-        if self.abbrListWidget.count() == 0:
-            self.removeButton.setEnabled(False)
-
-    def on_abbrListWidget_itemChanged(self, item):
-        if EMPTY_FIELD_REGEX.match(item.text()):
-            row = self.abbrListWidget.row(item)
-            self.abbrListWidget.takeItem(row)
-            del item
-
-        if self.abbrListWidget.count() == 0:
-            self.removeButton.setEnabled(False)
-
-    def on_abbrListWidget_itemDoubleClicked(self, item):
-        self.abbrListWidget.editItem(item)
-
-    def on_ignoreCaseCheckbox_stateChanged(self, state):
-        if not self.ignoreCaseCheckbox.isChecked():
-            self.matchCaseCheckbox.setChecked(False)
-
-    def on_matchCaseCheckbox_stateChanged(self, state):
-        if self.matchCaseCheckbox.isChecked():
-            self.ignoreCaseCheckbox.setChecked(True)
-
-    def on_immediateCheckbox_stateChanged(self, state):
-        if self.immediateCheckbox.isChecked():
-            self.omitTriggerCheckbox.setChecked(False)
-            self.omitTriggerCheckbox.setEnabled(False)
-            self.wordCharCombo.setEnabled(False)
-        else:
-            self.omitTriggerCheckbox.setEnabled(True)
-            self.wordCharCombo.setEnabled(True)
 
 
 class AbbrSettingsDialog(KDialog):
