+++ lib/autokey/qtui/resources/ui/abbrsettings.ui
@@ -1,0 +1,152 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ui version="4.0">
+ <class>Form</class>
+ <widget class="QWidget" name="Form">
+  <property name="geometry">
+   <rect>
+    <x>0</x>
+    <y>0</y>
+    <width>571</width>
+    <height>350</height>
+   </rect>
+  </property>
+  <property name="windowTitle">
+   <string>Form</string>
+  </property>
+  <layout class="QGridLayout" name="gridLayout">
+   <item row="2" column="5">
+    <widget class="QComboBox" name="wordCharCombo">
+     <property name="editable">
+      <bool>true</bool>
+     </property>
+    </widget>
+   </item>
+   <item row="6" column="3" colspan="4">
+    <widget class="QCheckBox" name="ignoreCaseCheckbox">
+     <property name="text">
+      <string>Ignore case of typed abbreviation</string>
+     </property>
+    </widget>
+   </item>
+   <item row="3" column="3" colspan="4">
+    <widget class="QCheckBox" name="removeTypedCheckbox">
+     <property name="text">
+      <string>Remove typed abbreviation</string>
+     </property>
+    </widget>
+   </item>
+   <item row="2" column="6">
+    <spacer name="horizontalSpacer">
+     <property name="orientation">
+      <enum>Qt::Horizontal</enum>
+     </property>
+     <property name="sizeHint" stdset="0">
+      <size>
+       <width>40</width>
+       <height>20</height>
+      </size>
+     </property>
+    </spacer>
+   </item>
+   <item row="2" column="3" colspan="2">
+    <widget class="QLabel" name="triggerOnLabel">
+     <property name="text">
+      <string>Trigger on:</string>
+     </property>
+    </widget>
+   </item>
+   <item row="5" column="3" colspan="4">
+    <widget class="QCheckBox" name="matchCaseCheckbox">
+     <property name="text">
+      <string>Match phrase case to typed abbreviation</string>
+     </property>
+    </widget>
+   </item>
+   <item row="8" column="3" colspan="4">
+    <widget class="QCheckBox" name="immediateCheckbox">
+     <property name="text">
+      <string>Trigger immediately (don't require a trigger character)</string>
+     </property>
+    </widget>
+   </item>
+   <item row="4" column="3" colspan="4">
+    <widget class="QCheckBox" name="omitTriggerCheckbox">
+     <property name="text">
+      <string>Omit trigger character</string>
+     </property>
+    </widget>
+   </item>
+   <item row="7" column="3" colspan="4">
+    <widget class="QCheckBox" name="triggerInsideCheckbox">
+     <property name="text">
+      <string>Trigger when typed as part of a word</string>
+     </property>
+    </widget>
+   </item>
+   <item row="2" column="0" rowspan="8" colspan="3">
+    <layout class="QGridLayout" name="gridLayout_2">
+     <item row="0" column="0" colspan="2">
+      <widget class="QListWidget" name="abbrListWidget">
+       <property name="sizePolicy">
+        <sizepolicy hsizetype="Ignored" vsizetype="Expanding">
+         <horstretch>0</horstretch>
+         <verstretch>0</verstretch>
+        </sizepolicy>
+       </property>
+       <property name="alternatingRowColors">
+        <bool>true</bool>
+       </property>
+       <property name="uniformItemSizes">
+        <bool>true</bool>
+       </property>
+      </widget>
+     </item>
+     <item row="1" column="1">
+      <widget class="QPushButton" name="removeButton">
+       <property name="text">
+        <string/>
+       </property>
+       <property name="icon">
+        <iconset theme="list-remove">
+         <normaloff>.</normaloff>.</iconset>
+       </property>
+      </widget>
+     </item>
+     <item row="1" column="0">
+      <widget class="QPushButton" name="addButton">
+       <property name="icon">
+        <iconset theme="list-add">
+         <normaloff>.</normaloff>.</iconset>
+       </property>
+       <property name="flat">
+        <bool>false</bool>
+       </property>
+      </widget>
+     </item>
+    </layout>
+   </item>
+   <item row="10" column="0" colspan="7">
+    <widget class="Line" name="line">
+     <property name="orientation">
+      <enum>Qt::Horizontal</enum>
+     </property>
+    </widget>
+   </item>
+   <item row="9" column="5">
+    <spacer name="verticalSpacer">
+     <property name="orientation">
+      <enum>Qt::Vertical</enum>
+     </property>
+     <property name="sizeHint" stdset="0">
+      <size>
+       <width>20</width>
+       <height>40</height>
+      </size>
+     </property>
+    </spacer>
+   </item>
+  </layout>
+ </widget>
+ <resources/>
+ <connections/>
+</ui>
